requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
scapy==2.5.0
geoip2==5.1.0
celery==5.3.4
redis==5.0.1
psutil==5.9.5
python-dotenv==1.0.0
pytest==7.4.3
coverage==7.3.2
gunicorn==21.2.0
